curalink-backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
motor==3.3.2
beanie==1.24.0
pymongo==4.6.1
cryptography==41.0.7
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
requests==2.31.0
websockets==12.0
aiofiles==23.2.1
biopython==1.83
